+++ src/declarations.d.ts
@@ -1,3 +1,4 @@
 declare module 'react-cursor-position';
 declare module 'color-sort';
 declare module 'react-easy-edit';
+declare module 'date-format';