+++ src/server/handlers/login.ts
@@ -1,0 +1,10 @@
+import { io } from 'fullstack-system';
+
+export default function (socket: SocketIO.Socket) {
+  io.sockets.emit('player-count', Object.keys(io.sockets.sockets).length);
+
+  socket.on('disconnect', () => {
+    // Update th player count for everyone else.
+    io.sockets.emit('player-count', Object.keys(io.sockets.sockets).length);
+  });
+}